c2-server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
redis==5.0.1
gunicorn==21.2.0
cryptography==41.0.7
